clean html extractor
