+++ airflow/dags/ingestion/ask-astro-load-airflow-docs.py
@@ -1,6 +1,7 @@
 import os
 from datetime import datetime
 
+import pandas as pd
 from include.utils.slack import send_failure_notification
 
 from airflow.decorators import dag, task
@@ -19,6 +20,15 @@ default_args = {"retries": 3, "retry_delay": 30}
 schedule_interval = "0 5 * * *" if ask_astro_env == "prod" else None
 
 
+@task
+def split_docs(urls, chunk_size=100) -> list[list[pd.Dataframe]]:
+    from include.tasks import split
+    from include.tasks.extract.utils.html_utils import urls_to_dataframe
+
+    chunked_urls = split.split_list(list(urls), chunk_size=chunk_size)
+    return [[urls_to_dataframe(chunk_url)] for chunk_url in chunked_urls]
+
+
 @dag(
     schedule_interval=schedule_interval,
     start_date=datetime(2023, 9, 27),
@@ -35,12 +45,9 @@ def ask_astro_load_airflow_docs():
     data from a point-in-time data capture. By using the upsert logic of the weaviate_import decorator
     any existing documents that have been updated will be removed and re-added.
     """
-    from include.tasks import split
     from include.tasks.extract import airflow_docs
 
     extracted_airflow_docs = task(airflow_docs.extract_airflow_docs)(docs_base_url=airflow_docs_base_url)
-
-    split_md_docs = task(split.split_html).expand(dfs=[extracted_airflow_docs])
 
     _import_data = WeaviateDocumentIngestOperator.partial(
         class_name=WEAVIATE_CLASS,
@@ -50,7 +57,7 @@ def ask_astro_load_airflow_docs():
         verbose=True,
         conn_id=_WEAVIATE_CONN_ID,
         task_id="WeaviateDocumentIngestOperator",
-    ).expand(input_data=[split_md_docs])
+    ).expand(input_data=split_docs(extracted_airflow_docs, chunk_size=100))
 
 
 ask_astro_load_airflow_docs()
