+++ airflow/include/tasks/extract/astro_forum_docs.py
@@ -7,7 +7,8 @@ import pandas as pd
 import pytz
 import requests
 from bs4 import BeautifulSoup
-from weaviate.util import generate_uuid5
+
+from include.tasks.extract.utils.html_utils import fetch_page_content, urls_to_dataframe
 
 cutoff_date = datetime(2022, 1, 1, tzinfo=pytz.UTC)
 
@@ -77,98 +78,12 @@ def get_cutoff_questions(forum_url: str) -> set[str]:
         page_url = f"{base_url}{page_number}"
         logger.info(page_url)
         page_number = page_number + 1
-        html_content = requests.get(page_url).content
+        html_content = fetch_page_content(page_url)
         questions_urls = get_questions_urls(html_content)
         if not questions_urls:  # reached at the end of page
             return set(all_valid_url)
         filter_questions_urls = filter_cutoff_questions(questions_urls)
         all_valid_url.extend(filter_questions_urls)
-
-
-def truncate_tokens(text: str, encoding_name: str, max_length: int = 8192) -> str:
-    """
-    Truncates a text string based on the maximum number of tokens.
-
-    param string (str): The input text string to be truncated.
-    param encoding_name (str): The name of the encoding model.
-    param max_length (int): The maximum number of tokens allowed. Default is 8192.
-    """
-    import tiktoken
-
-    try:
-        encoding = tiktoken.encoding_for_model(encoding_name)
-    except ValueError as e:
-        raise ValueError(f"Invalid encoding_name: {e}")
-
-    encoded_string = encoding.encode(text)
-    num_tokens = len(encoded_string)
-
-    if num_tokens > max_length:
-        text = encoding.decode(encoded_string[:max_length])
-
-    return text
-
-
-def clean_content(row_content: str) -> str | None:
-    """
-    Cleans and extracts text content from HTML.
-
-    param row_content (str): The HTML content to be cleaned.
-    """
-    soup = BeautifulSoup(row_content, "html.parser").find("body")
-
-    if soup is None:
-        return
-    # Remove script and style tags
-    for script_or_style in soup(["script", "style"]):
-        script_or_style.extract()
-
-    # Get text and handle whitespaces
-    text = " ".join(soup.stripped_strings)
-    # Need to truncate because in some cases the token size
-    # exceeding the max token size. Better solution can be get summary and ingest it.
-    return truncate_tokens(text, "gpt-3.5-turbo", 7692)
-
-
-def fetch_url_content(url) -> str | None:
-    """
-    Fetches the content of a URL.
-
-    param url (str): The URL to fetch content from.
-    """
-    try:
-        response = requests.get(url)
-        response.raise_for_status()  # Raise an HTTPError for bad responses
-        return response.content
-    except requests.RequestException:
-        logger.info("Error fetching content for %s: %s", url, url)
-        return None
-
-
-def process_url(url: str, doc_source: str = "") -> dict | None:
-    """
-    Process a URL by fetching its content, cleaning it, and generating a unique identifier (SHA) based on the cleaned content.
-
-    param url (str): The URL to be processed.
-    """
-    content = fetch_url_content(url)
-    if content is not None:
-        cleaned_content = clean_content(content)
-        sha = generate_uuid5(cleaned_content)
-        return {"docSource": doc_source, "sha": sha, "content": cleaned_content, "docLink": url}
-
-
-def url_to_df(urls: set[str], doc_source: str = "") -> pd.DataFrame:
-    """
-    Create a DataFrame from a list of URLs by processing each URL and organizing the results.
-
-    param urls (list): A list of URLs to be processed.
-    """
-    df_data = [process_url(url, doc_source) for url in urls]
-    df_data = [entry for entry in df_data if entry is not None]  # Remove failed entries
-    df = pd.DataFrame(df_data)
-    df = df[["docSource", "sha", "content", "docLink"]]  # Reorder columns if needed
-    return df
 
 
 def get_forum_df() -> list[pd.DataFrame]:
@@ -177,5 +92,5 @@ def get_forum_df() -> list[pd.DataFrame]:
     """
     questions_links = get_cutoff_questions("https://forum.astronomer.io/latest")
     logger.info(questions_links)
-    df = url_to_df(questions_links, "astro-forum")
+    df = urls_to_dataframe(questions_links, "astro-forum")
     return [df]
